DCT complete

+++ headers/dct.hpp
@@ -9,5 +9,5 @@ public:
 
   void performDCT(vector<unsigned char> &imageSubSample, unsigned int width,
                   unsigned int height);
-  void dctOn8x8(unsigned int i, unsigned int j, vector<vector<int>> &maxtix);
+  void dctOn8x8(unsigned int i, unsigned int j, vector<vector<float>> &maxtix);
 };
